Samples collected in March and June 2016

Samples collected in March and June 2016

Usage

water_samples

Format

A tibble with 157 rows and 80 variables

localization

Name of the localization where the sample was taken.

geology

Composition of the ground.

utm_x

Geospatial data of the water sampling locations. The geographic coordinate system 'Arc 1960 / UTM zone 37S' (EPSG:21037) which is used for the areas of Kenya and Tanzania - south of equator and east of 36°E.

utm_y

Geospatial data of the water sampling locations. The geographic coordinate system 'Arc 1960 / UTM zone 37S' (EPSG:21037) which is used for the areas of Kenya and Tanzania - south of equator and east of 36°E.

date

Date the sample was taken.

conductivity

Conductivity of the sample in (µS/cm)

T_avg

Average ambient temperature at the time of the sampling

pH

Acidity/basicity of the sample using the pH value.

TOC

Total organic carbon (TOC) is an analytical parameter representing the concentration of organic carbon in a sample.

alkalinity

Alkalinity of the sample as mg of bicarbonate (HCO3) per liter (mg_HCO3/L).

DO

Dissolved oxygen (DO) levels in environmental water depend on the physiochemical and biochemical activities in water body and it is an important useful in pollution and waste treatment process control.

ORP

Oxidation reduction potential (ORP) in (mV).

eH

Redox potential (eH) in (mV).

NH4

Ammonium concentration in (mg/L).

Cl

Chlorine concentration in (mg/L)

SO4

Sulfate concentration in (mg/L)

NO3

Nitrate concentration in (mg/L)

PO4

Phosphate concentration in (mg/L)

Br

Bromine concentration in (mg/L)

F

Fluorine concentration in (mg/L)

Ca

Calcium concentration in (mg/L)

Mg

Magnesium concentration in (mg/L)

Na

Sodium concentration in (mg/L)

K

Potassium concentration in (mg/L)

Fe

Iron concentration in (mg/L)

Si

Silicon concentration in (mg/L)

Al

Aluminum concentration in (mg/L)

S

Sulfur concentration in (mg/L)

P

Phosphorus concentration in (mg/L)

Li

Lithium concentration in parts per billion (ppb).

Be

Beryllium concentration in parts per billion (ppb).

B

Boron concentration in parts per billion (ppb).

Sc

Scandium concentration in parts per billion (ppb).

Ti

Titanium concentration in parts per billion (ppb).

V

Vanadium concentration in parts per billion (ppb).

Cr

Chromium concentration in parts per billion (ppb).

Mn

Manganese concentration in parts per billion (ppb).

Co

Cobalt concentration in parts per billion (ppb).

Ni

Nickel concentration in parts per billion (ppb).

Cu

Copper concentration in parts per billion (ppb).

Zn

Zinc concentration in parts per billion (ppb).

Ga

Gallium concentration in parts per billion (ppb).

Ge

Germanium concentration in parts per billion (ppb).

As

Arsenic concentration in parts per billion (ppb).

Se

Selenium concentration in parts per billion (ppb).

Rb

Rubidium concentration in parts per billion (ppb).

Sr

Strontium concentration in parts per billion (ppb).

Y

Yttrium concentration in parts per billion (ppb).

Zr

Zirconium concentration in parts per billion (ppb).

Nb

Niobium concentration in parts per billion (ppb).

Mo

Molybdenum concentration in parts per billion (ppb).

Cd

Cadmium concentration in parts per billion (ppb).

Sn

Tin concentration in parts per billion (ppb).

Sb

Antimony concentration in parts per billion (ppb).

Cs

Cesium concentration in parts per billion (ppb).

Ba

Barium concentration in parts per billion (ppb).

La

Lanthanum concentration in parts per billion (ppb).

Ce

Cerium concentration in parts per billion (ppb).

Pr

Praseodymium concentration in parts per billion (ppb).

Nd

Neodymium concentration in parts per billion (ppb).

Sm

Samarium concentration in parts per billion (ppb).

Eu

Europium concentration in parts per billion (ppb).

Gd

Gadolinium concentration in parts per billion (ppb).

Tb

Terbium concentration in parts per billion (ppb).

Dy

Dysprosium concentration in parts per billion (ppb).

Ho

Holmium concentration in parts per billion (ppb).

Er

Erbium concentration in parts per billion (ppb).

Tm

Thulium concentration in parts per billion (ppb).

Yb

Ytterbium concentration in parts per billion (ppb).

Lu

Lutetium concentration in parts per billion (ppb).

Hf

Hafnium concentration in parts per billion (ppb).

Ta

Tantalum concentration in parts per billion (ppb).

W

Wolfram concentration in parts per billion (ppb).

Tl

Thallium concentration in parts per billion (ppb).

Pb

Lead concentration in parts per billion (ppb).

Bi

Bismuth concentration in parts per billion (ppb).

Th

Thorium concentration in parts per billion (ppb).

U

Uranium concentration in parts per billion (ppb).

delta_O_18

The ratio of stable isotopes oxygen-18 (18O) and oxygen-16 (16O) as a measure of groundwater/mineral interactions.

delta_H_2

δ2H, or delta deuterium, is a measure of the relative abundance of deuterium (a stable isotope of hydrogen) in a sample, often used in hydrology and environmental science to trace the origin and movement of water.